requirements.txt
future==1.0.0
pynvml==12.0.0
prompt_toolkit==3.0.51
prometheus_client==0.21.1
annotated-types==0.7.0
markdown-it-py==3.0.0
pyJoules==0.5.1
codecarbon==3.0.2
experiment-impact-tracker==0.1.8
faiss-cpu==1.10.0
gensim==4.3.3
igraph==0.11.8
joblib==1.4.2
Levenshtein==0.27.1
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
peft==0.3.0
pillow==11.1.0
multiprocess==0.70.14
matplotlib==3.10.1
PyYAML==6.0.2
pyparsing==3.2.1
pyarrow==11.0.0
regex==2024.11.6
triton==3.4.0
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.13.2
yarl==1.20.1
yaspin==3.1.0

ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.5.1
scipy==1.13.1
threadpoolctl==3.5.0
tqdm==4.67.1
torch==1.13.1+cu117
torchvision==0.14.1+cu117
torchaudio==0.13.1

confluent-kafka==2.11.1


transformers==4.30.2
huggingface-hub==0.16.4
tokenizers==0.13.3 
sentence-transformers==2.2.2
datasets==2.14.5
evaluate==0.4.0
accelerate==0.20.3